Refactor and some new additions
Laying down the ground work for this library so there may be
some more restructures down the line.

Modules:
	- bencode: Encode/decode from bytes to python types
	- models: Contains data models w/ validation logic
	- tracker: Tracker client using http requests

Any models can be extetended with more fields, I just tried to
cover the simplest case first. The tracker module could use
some work as well, I know there is data in the response that I
haven't looked at yet. Also a tracker server implementation
would be a nice addition.

Some things for future work will be a 'peer' module for handling
p2p interactions. As well as a module for managing file pieces
as they are downloaded. And further down the line a cli tool.